requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
minio
python-jose[cryptography]
pytest
httpx
python-dotenv
pydantic_settings